+++ gDRplots/NEWS.md
@@ -1,3 +1,6 @@
+## gDRplots 0.0.83 - 2025-06-24
+* clean gene names in PRISM reports
+
 ## gDRplots 0.0.82 - 2025-06-06
 * update `pheatmap_qc` with standardized concentrations